requirements.txt
PyDrive
watchdog
oauth2client
google-api-python-client
google-auth
google-auth-httplib2